fix: extract FAQ into client component so page exports SEO metadata
The free-training-programs page had "use client" which blocked
metadata export. Extract the AccordionItem FAQ section into a
separate client component so the page can be a Server Component
with proper title, description, and canonical URL.
